+++ tests/test-bootstrap.php
@@ -1,7 +1,12 @@
 <?php
 use FatPanda\Illuminate\WordPress\TestCase;
 
+/**
+ * Test all bootstrapping
+ */
 class TestBootstrap extends TestCase {
+
+	protected $plugin = 'test-plugin';
 
 	/**
 	 * Make sure that our test framework is in place