+++ tests/test-routing.php
@@ -1,0 +1,17 @@
+<?php
+use FatPanda\Illuminate\WordPress\TestCase;
+
+/**
+ * Test all routing features of the framework
+ */
+class TestRouting extends TestCase {
+
+	/**
+	 * Test a GET request through the REST API
+	 */
+	function testGet()
+	{
+			
+	}
+
+}